AI_Service/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
optuna
lightgbm
xgboost
catboost
scikit-learn
scipy
matplotlib 
seaborn
joblib
python-dotenv
pytest
httpx
